Parte 4 completada

+++ part4/Lista de blogs/controllers/blogs.js
@@ -1,31 +1,68 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
 
 //get all blogs
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+
   response.json(blogs)
+})
+
+//get a specific blog
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  response.json(blog)
 })
 
 //Post a blog
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id) {
+    return response.status(401).json({ error: 'Token missing or invalid'})
+  }
+
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     author: body.author,
     title: body.title,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.json(savedBlog)
 })
 
 //Delete a blog
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+
+  if ( !blog.user ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  const user = await User.findById(decodedToken.id)
+  
+  if ( blog.user.toString() === user.id ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    response.status(400).json({error: "can't delete a blog is not yours"})
+  }
 })
 
 //Update a blog
